Fix assert_has_calls use in test
This is defined to accept a list of calls, not a single call. Presumably
a newer version of mock became stricter about this, or a change in
libvirt meant that an additional call was being made, which caused an
inadvertent change in test behaviour because of the call to
assert_has_calls not following the definition.