Fix dnsmasq lease inotify watcher
dnsmasq seems to have changed to renaming a new file in to the leases
file. As we were watching for modifications to this file, file updates
were being missed, so the "uvt-kvm wait" command was failing to see the
changes and eventually timing out.

Now we additionally watch for renames in to the expected path to pick up
on this (better) style of updating the file.