Pull in d/changelog from the Ubuntu archive
This is an update from 0~bzr99-0ubuntu2 that I uploaded the other day.

The changes in that upload are already covered by either the master
branch (upstream code) or this branch (packaging).

This branch is not yet fully in sync with Ubuntu, but I am keeping it
close. I intend to align it fully in Yakkety+1 when I upload from here
for the first time.