d/control: update libvirt dependency
Post Xenial libvirt started to re-align with Debian packaging.
That affects the package names and content in those.
libvirt-bin in particular is only a transitional package which we want
to get rid of (LP: #1694159).

Currently uvtool-libvirt depended on it for two things:
1. to have a working libvirt service it can connect
2. to let users interact via virsh after uvt created a guest

Of these #1 is a real dependency - and while libvirt-daemon brings the
binaries it is only libvirt-daemon-system that does install it as a
service as people (and uvtool) expect it.

I first thought #2 would only be a recommends as most users want tools
like "virsh" to manage the uvtool created guests. But I realized that
while uvtool itself only connects via python-libvirt the maintainer
scripts of uvtool use virsh - therefore it has to be a dependency to
libvirt-clients as well.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>